Needless optimization
Unroll the loop once for the sake of demonstration:
By unrolling the loop once we mean that we do
twice the body/work of the previous loop per iteration
of the new loop. Hence, we iterate only half the times
which saves some cpu cycles, namely copying input back
to previous. In this case I even skipped the test that
tests if we're finished, assuming we need to loop a lot
that will increase the speed.

Note that normally we only talk about unrolling a loop
when that means that the compiler can do things in parallel,
really getting twice as fast. That is not the case here
because reachable2 is calculated from reachable1.
Also note that unrolling a loop is what the compiler does
already itself, when optimizing (if it's smart enough).
It would never skip the test, as we do here, though.

Whether or not the compiler will put Directions(1,1,1,1)
outside the loop depends on whether or not that constructor
is inlined: if it is not inlined then the constructor MIGHT
do something else every time (ie, have an internal counter)
and therefore has to be called every time. While when it
is inlined then the compiler will see that that is not the
case and probably put it outside the loop. If unsure, we
could put it outside the loop ourselves.