Revert loop unrolling.
The unrolling made too little sense.
This is the old code again, except that
I renamed 'input' to 'reachable'. The reason
for that is that it's only input after the first
line of the function, after that the variable
is updated to be what is reachable from that
when you're allowed to flow through 'space'.